fixed get_msg arg problem on server

+++ both/get_file_mmap.c
@@ -1,6 +1,6 @@
 #include "both.h"
 
-void		*get_file_mmap(char *filename, uint32_t *size)
+void		*get_file_mmap(char *filename, uint64_t *size)
 {
 	struct stat st;
 	void		*file;
